WIP: feat: expose `wgpu::Device::RECOMMENDED_MIN_STACK_SPACE_FOR_NATIVE_WGSL_COMPILATION`
TODO: I make a copy of the recommended stack size in `wgpu_types` here,
but I'm not sure it's necessary. I don't think we need to expose this
under any conditions besides those in which `naga` will also be exposed,
so we might just be able to keep the recommendation there.